Extend mirage start helper to allow additional options
This makes it so other modules can utilize the `setupApplication`
helper and provide their own mirage configuration that would extend
the core mirage configuration.
